Migrates from mezzo to freestyle-rpc style, license, etc.
