+++ .travis.yml
@@ -1,30 +1,44 @@
 language: scala
+
+scala:
+- 2.11.11
+- 2.12.2
+
 jdk:
 - oraclejdk8
-scala:
-- 2.12.1
-- 2.11.8
 
-#install:
-#- pip install --user codecov
-
-script:
-- sbt ++$TRAVIS_SCALA_VERSION validate
-#- sbt ++$TRAVIS_SCALA_VERSION coverage validate coverageReport
-#- codecov
-
-cache:
-  directories:
-  - $HOME/.sbt/0.13
-  - $HOME/.sbt/boot/scala*
-  - $HOME/.sbt/cache
-  - $HOME/.sbt/launchers
-  - $HOME/.ivy2
-  - $HOME/.coursier
 before_cache:
 - du -h -d 1 $HOME/.ivy2/
 - du -h -d 2 $HOME/.sbt/
 - du -h -d 4 $HOME/.coursier/
 - find $HOME/.sbt -name "*.lock" -type f -delete
 - find $HOME/.ivy2/cache -name "ivydata-*.properties" -type f -delete
-- find $HOME/.coursier/cache -name "*.lock" -type f -delete
+
+cache:
+  directories:
+  - $HOME/.sbt/cache
+  - $HOME/.sbt/0.13
+  - $HOME/.sbt/boot/
+  - $HOME/.sbt/boot/scala*
+  - $HOME/.sbt/launchers
+  - $HOME/.ivy2/cache
+  - $HOME/.ivy2
+
+before_install:
+- if [ "$TRAVIS_BRANCH" = "master" -a "$TRAVIS_PULL_REQUEST" = "false" ]; then
+    openssl aes-256-cbc -K $encrypted_80bb47bfd841_key -iv $encrypted_80bb47bfd841_iv -in secring.gpg.enc -out secring.gpg -d;
+  fi
+- export PATH=${PATH}:./vendor/bundle
+
+install:
+- rvm use 2.2.3 --install --fuzzy
+- gem update --system
+- gem install sass
+- gem install jekyll -v 3.4.3
+
+script:
+- sbt ++$TRAVIS_SCALA_VERSION orgScriptCI
+
+after_success:
+- bash <(curl -s https://codecov.io/bash) -t token_replace-me
+- sbt ++$TRAVIS_SCALA_VERSION orgAfterCISuccess
