+++ AUTHORS.md
@@ -1,0 +1,12 @@
+# Authors
+
+## Maintainers
+
+The maintainers of the project are:
+
+* 47 Degrees (twitter: @47deg) <[47degfreestyle](https://github.com/47degfreestyle)>
+
+## Contributors
+
+These are the people that have contributed to the freestyle-rpc project:
+
