+++ CONTRIBUTING.md
@@ -1,0 +1,27 @@
+# Contributing
+
+Discussion around Freestyle happens in the [Gitter channel](https://gitter.im/47deg/freestyle) as well as on
+[GitHub issues](https://github.com/47deg/freestyle/issues) and [pull requests](https://github.com/47deg/freestyle/pulls).
+
+Feel free to open an issue if you notice a bug, have an idea for a feature, or have a question about
+the code. Pull requests are also welcome.
+
+People are expected to follow the [Typelevel Code of Conduct](http://typelevel.org/conduct.html) when discussing Freestyle on the Github page, Gitter channel, or other venues.
+
+If you are being harassed, please contact one of [us](AUTHORS.md#maintainers) immediately so that we can support you. In case you cannot get in touch with us please write an email to [47 Degrees](mailto:hello@47deg.com).
+
+## How can I help?
+
+Freestyle follows a standard [fork and pull](https://help.github.com/articles/using-pull-requests/) model for contributions via GitHub pull requests.
+
+The process is simple:
+
+ 1. Find something you want to work on
+ 2. Let us know you are working on it via the Gitter channel or GitHub issues/pull requests
+ 3. Implement your contribution
+ 4. Write tests
+ 5. Update the documentation
+ 6. Submit pull request
+
+You will be automatically included in the [AUTHORS.md](AUTHORS.md#contributors) file as contributor in the next release.
+If you encounter any confusion or frustration during the contribution process, please create a GitHub issue and we'll do our best to improve the process.
