+++ NOTICE.md
@@ -1,0 +1,4 @@
+Freestyle
+Copyright (c) 2017 47 Degrees.  All rights reserved.
+
+Licensed under Apache License. See [LICENSE](LICENSE) for terms.
