+++ http/core/src/main/scala_2.11/package.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 47 Degrees, LLC. <http://www.47deg.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package freestyle
+package http
+package core
+
+package object core {
+  private[core] implicit class EitherCompatOps[A, B](val either: Either[A, B]) extends AnyVal {
+    def flatMap[AA >: A, C](f: B => Either[AA, C]): Either[AA, C] =
+      either.right.flatMap(f)
+    def map[AA >: A, C](f: B => C): Either[AA, C] =
+      either.right.map(f)
+  }
+}
