--- modules/http-akka/src/main/scala/mezzo/h2akka/package.scala
@@ -1,15 +1,0 @@
-/* -
- * Mezzo [http-akka]
- */
-
-package mezzo
-
-package object h2akka {
-  type AkkaHttpClient = Dehydrated {
-    type H = HydrateAkkaHttpClient[_]
-  }
-
-  type AkkaHttpRoutes = Dehydrated {
-    type H = HydrateAkkaHttpRoutes[_]
-  }
-}
