+++ project/plugins.sbt
@@ -1,7 +1,2 @@
-addSbtPlugin("de.heikoseeberger" % "sbt-header"      % "1.5.1")
-addSbtPlugin("org.tpolecat"      % "tut-plugin"      % "0.4.8")
-addSbtPlugin("com.eed3si9n"      % "sbt-unidoc"      % "0.3.3")
-addSbtPlugin("org.scoverage"     % "sbt-scoverage"   % "1.5.0")
-addSbtPlugin("com.fortysevendeg" % "sbt-microsites"  % "0.4.0")
-addSbtPlugin("com.jsuereth"      % "sbt-pgp"         % "1.0.0")
-addSbtPlugin("io.get-coursier"   % "sbt-coursier"    % "1.0.0-M15")
+resolvers ++= Seq(Resolver.sonatypeRepo("snapshots"), Resolver.sonatypeRepo("releases"))
+addSbtPlugin("io.frees" % "sbt-freestyle" % "0.0.1-SNAPSHOT")
