+++ version.sbt
@@ -1,0 +1,1 @@
+version in ThisBuild := "0.0.1-SNAPSHOT"
